Moved `ImageCommentsUIComposer` and collaborators `ImageCommentsLoaderPresentationAdapter` and `MainQueueDispatchDecorator` to production target
